+++ Cargo.toml
@@ -196,6 +196,7 @@ itoa = { version = "1", default-features = false }
 ryu = { version = "1", default-features = false }
 colored = { version = "3", default-features = false }
 tracing-rolling-file = "0.1"
+tracing-appender = { version = "0.2", default-features = false }
 
 [profile.release]
 debug = true
